services/talking_head/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

numpy==1.26.4
opencv-python-headless==4.10.0.84
imageio==2.35.1
imageio-ffmpeg==0.5.1

librosa==0.9.2
tqdm==4.66.4
Pillow==10.4.0

torch==2.3.1
torchaudio==2.3.1
